+++ src/soyokaze/commands/uiautomation/WindowUIElement.cpp
@@ -1,6 +1,5 @@
 #include "pch.h"
 #include "WindowUIElement.h"
-#include "commands/uiautomation/UIElementAliasMap.h"
 #include "utility/ScopeAttachThreadInput.h"
 #include "SharedHwnd.h"
 
@@ -138,89 +137,6 @@ bool WindowUIElement::CanFocus()
 	mElem->get_CachedIsKeyboardFocusable(&isFocusable);
 	return isFocusable != FALSE;
 }
-
-////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
-
-
-Win32MenuItemElement::Win32MenuItemElement()
-{
-}
-
-Win32MenuItemElement::Win32MenuItemElement(HWND hwnd, UINT id, const CString& name) : 
-	mHwnd(hwnd), mMenuId(id), mName(name)
-{
-}
-
-Win32MenuItemElement::~Win32MenuItemElement()
-{
-}
-
-
-CString Win32MenuItemElement::GetName()
-{
-	auto nameMap = UIElementAliasMap::GetInstance();
-
-	TCHAR clsName[256];
-	GetClassName(mHwnd, clsName, 256);
-
-	CString alias;
-	if (nameMap->GetAlias(clsName, mName, mMenuId, alias)) {
-		return alias;
-	}
-	else {
-		return mName;
-	}
-}
-
-CRect Win32MenuItemElement::GetRect()
-{
-	return CRect(0,0,0,0);
-}
-
-bool Win32MenuItemElement::Click()
-{
-	if (IsWindow(mHwnd) == FALSE) {
-		return false;
-	}
-
-	// 親ウインドウを前面に出す
-	ScopeAttachThreadInput scope;
-	SetForegroundWindow(mHwnd);
-
-	// メニューを選択する操作
-	SendMessage(mHwnd, WM_COMMAND, MAKEWPARAM(mMenuId, 0), 0);
-
-	return true;
-}
-
-bool Win32MenuItemElement::Focus()
-{
-	return false;
-}
-
-bool Win32MenuItemElement::CanFocus()
-{
-	return false;
-}
-
-
-uint32_t Win32MenuItemElement::AddRef()
-{
-	return InterlockedIncrement(&mRefCount);
-}
-
-uint32_t Win32MenuItemElement::Release()
-{
-	auto n = InterlockedDecrement(&mRefCount);
-	if (n == 0) {
-		delete this;
-	}
-	return n;
-}
-
-
 
 }}} // end of namespace launcherapp::commands::uiautomation